+++ internal/menubar/changelog.go
@@ -1,0 +1,3 @@
+package menubar
+
+var embeddedChangelog string